docs: added in-code documentation
