+++ components/ethers/include/firefly-rlp.h
@@ -9,7 +9,16 @@ extern "C" {
 #include <stdint.h>
 
 
-//typedef uint16_t RlpBuilderTag;
+/**
+ *  Recursive-Length Prefix (RLP) Encoder
+ *
+ *  RLP-encoding is used for various purposes in Ethereum, such as
+ *  serializing transactions, which is in turn used to create the
+ *  necessary hashes for signing.
+ *
+ *  See: https://ethereum.org/en/developers/docs/data-structures-and-encoding/rlp/
+ */
+
 
 typedef enum FfxRlpStatus {
     FfxRlpStatusOK = 0,
@@ -18,10 +27,12 @@ typedef enum FfxRlpStatus {
     FfxRlpStatusOverflow = -55
 } FfxRlpStatus;
 
+
 typedef struct FfxRlpBuilder {
     uint8_t *data;
     size_t offset, length;
 } FfxRlpBuilder;
+
 
 /**
  *  Initializes a new RLP builder.
@@ -38,14 +49,28 @@ void ffx_rlp_build(FfxRlpBuilder *builder, uint8_t *data, size_t length);
  */
 size_t ffx_rlp_finalize(FfxRlpBuilder *builder);
 
+
+/**
+ *  Append a Data.
+ */
 FfxRlpStatus ffx_rlp_appendData(FfxRlpBuilder *builder, uint8_t *data,
   size_t length);
+
+/**
+ *  Append a Data, with the contents of the a NULL-terminated string.
+ */
 FfxRlpStatus ffx_rlp_appendString(FfxRlpBuilder *builder, const char *data);
 
+/**
+ *  Append an Array, where the next %%count%% Items added will be
+ *  added to this array.
+ */
 FfxRlpStatus ffx_rlp_appendArray(FfxRlpBuilder *builder, size_t count);
 
-//RlpStatus rlp_appendArrayMutable(RlpBuilder *builder, RlpBuilderTag *tag);
 
+// @TODO: Future API?
+//typedef uint16_t RlpBuilderTag;
+//RlpStatus rlp_appendArrayMutable(RlpBuilder *builder, RlpBuilderTag *tag);
 //void rlp_adjustCount(RlpBuilder *builder, RlpBuilderTag tag, uint16_t count);
